+++ packages/language-server/src/documents/DocumentCache.ts
@@ -66,7 +66,7 @@ export class DocumentCache {
         if (variableName === '_self') return document;
 
         const scopedImports = pos
-            ? document.getScopeAt(pos)?.symbols.imports
+            ? document.getScopeAt(pos)?.imports
             : document.locals.imports
 
         const twigImport = scopedImports?.find(imp => imp.name === variableName);