+++ packages/language-server/src/server.ts
@@ -23,6 +23,8 @@ import {
 import { initializeParser } from './utils/parser';
 import { BracketSpacesInsertionProvider } from './autoInsertions/BracketSpacesInsertionProvider';
 import { InlayHintProvider } from './inlayHints/InlayHintProvider';
+import { ReferenceProvider } from './references/ReferenceProvider';
+import { RenameProvider } from './references/RenameProvider';
 
 export class Server {
     readonly documents = new TextDocuments(TextDocument);
@@ -34,6 +36,8 @@ export class Server {
     bracketSpacesInsertionProvider!: BracketSpacesInsertionProvider;
     inlayHintProvider!: InlayHintProvider;
     signatureHelpProvider!: SignatureHelpProvider;
+    referenceProvider!: ReferenceProvider;
+    renameProvider!: RenameProvider;
     diagnosticProvider: DiagnosticProvider;
 
     constructor(connection: Connection) {
@@ -51,6 +55,8 @@ export class Server {
             new SymbolProvider(connection, documentCache);
             new HoverProvider(connection, documentCache);
             this.signatureHelpProvider = new SignatureHelpProvider(connection, documentCache);
+            this.referenceProvider = new ReferenceProvider(connection, documentCache);
+            this.renameProvider = new RenameProvider(connection, documentCache);
             this.definitionProvider = new DefinitionProvider(
                 connection,
                 documentCache,
@@ -84,6 +90,10 @@ export class Server {
                     full: true,
                 },
                 inlayHintProvider: true,
+                referencesProvider: true,
+                renameProvider: {
+                    prepareProvider: true,
+                },
                 executeCommandProvider: {
                     commands: [
                         `${Command.IsInsideHtmlRegion}(${this.workspaceFolder.uri})`,